+++ src/Core/ProductBundle/Controller/MediaController.php
@@ -111,18 +111,30 @@ class MediaController extends TranslateController
 
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getManager();
-            if($type == 'image')
+            $testEntity = $em->getRepository('CoreMediaBundle:Media')->findOneByHash($entity->getHash());
+            if($testEntity !== null)
             {
-                $imageOptions = $this->container->getParameter('images');
-                $opts = array();
-                foreach($this->container->getParameter('product.images') as $val)
-                {
-                    $opts[$val] = $imageOptions[$val];
-                }
-                $entity->setOptions($opts);
+                $testEntity->setUsedCount($testEntity->getUsedCount() + 1);
+                $em->persist($testEntity);
+                $em->flush();
+                $entity = $testEntity;
             }
-            $em->persist($entity);
-            $em->flush();
+            else   
+            {
+                if($type == 'image')
+                {
+                    $imageOptions = $this->container->getParameter('images');
+                    $opts = array();
+                    foreach($this->container->getParameter('product.images') as $val)
+                    {
+                        $opts[$val] = $imageOptions[$val];
+                    }
+                    $entity->setOptions($opts);
+                }
+                $em->persist($entity);
+                $em->flush();
+                $this->saveTranslations($entity, $cultures);
+            }
             $productEntity = $em->getRepository('CoreProductBundle:Product')->find($product);
             if($productEntity != null)
             {
@@ -130,7 +142,6 @@ class MediaController extends TranslateController
                 $em->persist($productMedia);
                 $em->flush();
             }
-            $this->saveTranslations($entity, $cultures);
             return $this->redirect($this->generateUrl('product_media', array('category' => $category, 'product' => $product)));
             
         }
@@ -283,12 +294,29 @@ class MediaController extends TranslateController
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find ' . $type . ' entity.');
             }
-            if($type == 'image')
+            $productEntity = $em->getRepository('CoreProductBundle:Product')->find($product);
+            if($productEntity != null)
             {
-                $imageOptions = $this->container->getParameter('images');
-                $entity->setOptions($imageOptions);
+                $productMedia = $productEntity->getProductMedia($entity->getId());
+                if($productMedia)
+                {
+                    $em->remove($productMedia);
+                }
             }
-            $em->remove($entity);
+            $entity->setUsedCount($entity->getUsedCount() - 1);
+            if($entity->getUsedCount() == 0)
+            {
+                if($entity->getType() == 'image')
+                {
+                    $imageOptions = $this->container->getParameter('images');
+                    $entity->setOptions($imageOptions);
+                }
+                $em->remove($entity);
+            }
+            else
+            {
+                $em->persist($entity);
+            }
             $em->flush();
         }
